Added source code folder and organized
