--- 395project2019/.gitignore
@@ -1,1 +1,0 @@
-
