--- 395project2019/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=36856aac
-build.xml.script.CRC32=d9f0e7b0
-build.xml.stylesheet.CRC32=8064a381@1.80.1.48
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=36856aac
-nbproject/build-impl.xml.script.CRC32=4add7c6a
-nbproject/build-impl.xml.stylesheet.CRC32=830a3534@1.80.1.48
